requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
openai==0.28.1
Werkzeug==2.3.7
Jinja2==3.1.2
reportlab==4.0.4
openpyxl==3.1.2
python-docx==0.8.11
lxml==4.9.3
Pillow==10.0.1
pyreportjasper==2.1.2
pyjasper==0.41
